refactor: 스프링 시큐리티 삭제
