+++ user-service/src/main/java/com/palja/user_service/application/command/LoginUserCommand.java
@@ -1,0 +1,12 @@
+package com.palja.user_service.application.command;
+
+import lombok.Builder;
+
+@Builder
+public record LoginUserCommand(
+
+	String loginId,
+	String password
+
+) {
+}
