+++ user-service/src/main/java/com/palja/user_service/application/dto/response/TokenRes.java
@@ -1,0 +1,14 @@
+package com.palja.user_service.application.dto.response;
+
+import lombok.Builder;
+import lombok.Getter;
+
+@Getter
+@Builder
+public class TokenRes {
+
+	private String accessToken;
+	private String refreshToken;
+	private long refreshKeyExpirationTime;
+
+}
