--- user-service/src/main/java/com/palja/user_service/infrastructure/security/dto/request/LoginUserReq.java
@@ -1,14 +1,0 @@
-package com.palja.user_service.infrastructure.security.dto.request;
-
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class LoginUserReq {
-
-	private String loginId;
-	private String password;
-
-}
