+++ user-service/src/main/java/com/palja/user_service/presentation/controller/AuthController.java
@@ -1,20 +1,28 @@
 package com.palja.user_service.presentation.controller;
 
+import java.net.URLEncoder;
+import java.nio.charset.StandardCharsets;
+
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseCookie;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.CookieValue;
 import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.palja.common.auditor.AuditorContext;
 import com.palja.common.response.ApiResponse;
+import com.palja.user_service.application.command.LoginUserCommand;
+import com.palja.user_service.application.dto.response.TokenRes;
 import com.palja.user_service.application.service.AuthService;
+import com.palja.user_service.presentation.dto.request.LoginUserReq;
 
 import jakarta.servlet.http.HttpServletResponse;
+import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 
 @RestController
@@ -23,6 +31,22 @@ import lombok.RequiredArgsConstructor;
 public class AuthController {
 
 	private final AuthService authService;
+
+	@PostMapping("/login")
+	public ResponseEntity<ApiResponse<Void>> login(@Valid @RequestBody LoginUserReq requestDto, HttpServletResponse response) {
+		LoginUserCommand command = LoginUserReq.of(requestDto);
+		TokenRes tokenResponse = authService.login(command);
+
+		String accessToken = tokenResponse.getAccessToken();
+		addAccessTokenToHeader(response, accessToken);
+
+		String refreshToken = tokenResponse.getRefreshToken();
+		String encodedRefreshToken = URLEncoder.encode(refreshToken, StandardCharsets.UTF_8).replace("\\+", "%20");
+		long refreshKeyExpirationTime = tokenResponse.getRefreshKeyExpirationTime();
+		addRefreshTokenToCookie(response, encodedRefreshToken, refreshKeyExpirationTime);
+
+		return ResponseEntity.status(HttpStatus.OK).body(ApiResponse.success("로그인 되었습니다."));
+	}
 
 	@PostMapping("/refresh")
 	public ResponseEntity<ApiResponse<Void>> refresh(
@@ -43,7 +67,7 @@ public class AuthController {
 		String loginId = AuditorContext.get().getLoginId();
 
 		authService.logout(loginId, accessToken);
-		expireRefreshTokenToCookie(response);
+		addRefreshTokenToCookie(response, "", 0);
 
 		return ResponseEntity.status(HttpStatus.OK).body(ApiResponse.success("로그아웃 되었습니다."));
 	}
@@ -52,10 +76,13 @@ public class AuthController {
 		response.setHeader("Authorization", accessToken);
 	}
 
-	private void expireRefreshTokenToCookie(HttpServletResponse response) {
+	private void addRefreshTokenToCookie(HttpServletResponse response, String refreshToken, long maxAge) {
 		ResponseCookie cookie = ResponseCookie
-			.from("refresh_token")
-			.maxAge(0)
+			.from("refresh_token", refreshToken)
+			.path("/")
+			.httpOnly(true)
+			.secure(false)
+			.maxAge(maxAge)
 			.build();
 		response.addHeader(HttpHeaders.SET_COOKIE, cookie.toString());
 	}
